chatbotreforma/requirements.txt
fastapi==0.95.2
uvicorn[standard]==0.22.0
httpx==0.24.1
mysql-connector-python==8.1.0
python-dotenv==1.0.0
openai==0.27.0
psycopg2-binary==2.9.7
tqdm==4.65.0
python-multipart==0.0.20